feat(upload links): implement deactivation endpoint with error handling
In this commit, we query the database to find the upload link by ID. If the upload link does not exist a 404 error is returned,
otherwise the upload link will update the active status to false.

The response is returned with the full updated link object.

`AuthenticatedRequest` is used as the request type to access user

Refs: FF-33